% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colors.R
\docType{data}
\name{Board_colors}
\alias{Board_colors}
\title{These function loads the Board Color palettes which includes:}
\format{An object of class \code{character} of length 7.}
\usage{
Board_colors
}
\description{
Board_cols():
       turquoise, orange, yellow, burgundy, blue, grey, green
}
\details{
Board_pal():
         main  = blue, green, yellow
         cool  = blue, green
         hot   = yellow, orange, burgundy),
         mixed = blue, green, yellow, orange, burgundy),
         grey  = light grey, dark grey)

Board_scale_color()

Board_scale_fill()

#Use https://coolors.co/ to get color tints
}
\keyword{datasets}
