% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colors.R
\name{Board_scale_color}
\alias{Board_scale_color}
\title{Color scale constructor for drsimonj colors}
\usage{
Board_scale_color(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in drsimonj_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for drsimonj colors
}
