% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle.R
\name{circle}
\alias{circle}
\title{graphs and calculates area and perimeter of a circle}
\usage{
circle(x)
}
\arguments{
\item{x}{a numeric vector containing positive value(s)}
}
\description{
graphs and calculates area and perimeter of a circle
}
