% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_ngon.R
\name{coords_ngon}
\alias{coords_ngon}
\title{creates a tibble with the an entry for the radius, number of sides,
x and y coordinates, perimeter, and area of shape}
\usage{
coords_ngon(r, n)
}
\arguments{
\item{r}{a numeric input containing a positive integer}

\item{n}{a numeric input containing a positive integer}
}
\value{
a dataframe containining the radius \code{r}, number of side \code{n}
}
\description{
creates a tibble with the an entry for the radius, number of sides,
x and y coordinates, perimeter, and area of shape
}
\examples{
ngon(3,4)
ngon (102,40)
}
