% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_ngon.R
\name{draw_ngon}
\alias{draw_ngon}
\title{graphs and calculates the area and perimeter of a square}
\usage{
draw_ngon(ngon)
}
\arguments{
\item{ngon}{a tibble output from the function ngon}
}
\value{
a ggplot with the plot of the n-gon shape with radius \code{r} as defined in ngon
}
\description{
graphs and calculates the area and perimeter of a square
}
\examples{
draw_ngon(ngon(2,4))
}
