% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_problems.R
\name{generate_problems}
\alias{generate_problems}
\title{generates a table of practice problems for students}
\usage{
generate_problems(type, lower = 1, upper, repeats = 1)
}
\arguments{
\item{type}{a string; must be multiple, divide, add, or subtract}

\item{lower}{lower bound for generating random numbers; default value of 1}

\item{upper}{upper bound for generating random numbers}

\item{repeats}{number of problems to be generated; default is one}
}
\description{
generates a table of practice problems for students
}
