% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_vis.R
\name{graph_my_data}
\alias{graph_my_data}
\alias{graph_my_data.numeric}
\alias{graph_my_data.data.frame}
\title{graphs a given set of data}
\usage{
graph_my_data(x, y, color, ind_var, dep_var, title, data)

\method{graph_my_data}{numeric}(x, y, color, ind_var, dep_var, title,
  data = NULL)

\method{graph_my_data}{data.frame}(x = NULL, y = NULL, color, ind_var,
  dep_var, title, data)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}

\item{color}{a string; must be a color recognized by ggplot}

\item{ind_var}{a string of the variable name to be assessed; if using \code{data}, the vector of this variable name must have numerical values}

\item{dep_var}{a string of the variable name to be assessed; if using \code{data}, the vector of this variable name must have numerical values}

\item{title}{a string}

\item{data}{default value is NULL, so that if manually entering data as x and y vectors this is not involved}
}
\description{
graphs a given set of data
}
