% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ngon.R
\name{map_ngon}
\alias{map_ngon}
\title{creates a tibble with a new row for each of the entries of the vector
radius, \code{r} or the vector number of sides \code{n}
finding the respective x and y coordinates, perimeter, and area of each entry}
\usage{
map_ngon(math_ngon)
}
\arguments{
\item{math_ngon}{a tibble created from the function coords_ngon}
}
\value{
a tible containining the radius \code{r}, number of side \code{n},
area and perimeter of each the entries
}
\description{
creates a tibble with a new row for each of the entries of the vector
radius, \code{r} or the vector number of sides \code{n}
finding the respective x and y coordinates, perimeter, and area of each entry
}
\examples{
map_ngon(ngon(2:3,4))
map_ngon(ngon (2,4:5))
}
