% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngon.R
\name{ngon}
\alias{ngon}
\title{graphs and calculates the area and perimeter of a square}
\usage{
ngon(r, n)
}
\arguments{
\item{r}{a numeric input containing a positive integer}

\item{n}{a numeric input containing a positive integer}
}
\value{
a dataframe containining the radius, number of sides, area, perimeter,
and a graph of the n-gon with radius \code{r} and sides \code{s}
}
\description{
graphs and calculates the area and perimeter of a square
}
\examples{
ngon(3,4)
ngon (102,40)
}
