% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{graphs and calculates the area and perimeter of a square}
\usage{
square(x)
}
\arguments{
\item{x}{a numeric vector containing a positive integer}
}
\value{
a dataframe containining the length, width, area, perimeter,
and a graph of the square with length and width \code{x}
}
\description{
graphs and calculates the area and perimeter of a square
}
\examples{
square(3)
square (2:4)
}
