% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_ngon.R
\name{draw_ngon}
\alias{draw_ngon}
\title{graphs  the area and perimeter of an n-gon}
\usage{
draw_ngon(ngon_coord2)
}
\arguments{
\item{ngon_coord2}{a tibble output from the function map_ngon}
}
\value{
a ggplot with the plot of the n-gon shape with radius \code{r}
as defined in ngon
}
\description{
graphs  the area and perimeter of an n-gon
}
\examples{
draw_ngon(map_ngon(ngon(2,4)))
}
