% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngon.R
\name{ngon}
\alias{ngon}
\title{creates a tibble with the an entry for the radis and number of sides}
\usage{
ngon(r, n)
}
\arguments{
\item{r}{a numeric input containing a positive integer}

\item{n}{a numeric input containing a positive integer}
}
\value{
a dataframe containining the radius \code{r}, number of side \code{n}
}
\description{
creates a tibble with the an entry for the radis and number of sides
}
\examples{
ngon(3,4)
ngon (102,40)
}
